1.Update SessionClient class 增加Cookies读写 etc. 2.更新 网络连接参数 3.修正 WCosplay站点默认限制8张图 4.更新 MoeLoaderDelta主程序所需支持库为Net Framework 4.5.1 5.增加 RabbitBox 为支持用户登录站点做准备

+++ MoeLoaderDelta/Control/DownloadControl.xaml.cs
@@ -514,7 +514,7 @@ namespace MoeLoaderDelta
         {
             if (downloadItems.Count > 0)
             {
-                double percent = (double)(downloadItems.Count - numLeft - webs.Count) / (double)downloadItems.Count * 100.0;
+                double percent = (downloadItems.Count - numLeft - webs.Count) / (double)downloadItems.Count * 100.0;
 
                 Win7TaskBar.ChangeProcessValue(MainWindow.Hwnd, (uint)percent);
